requirements.txt
django
gunicorn
python-telegram-bot
nest_asyncio
requests
python-dotenv
